refactor: split array logic
